+++ Changelog.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- support for using a different config file name, not only `.env.json`
+- support for expanding environment variables inside of the config file
+
 ## [1.1.0] - 2022-09-30
 ### Added